\name{icemelt-package}
\alias{icemelt-package}
\alias{icemelt}
\docType{package}
\title{
\packageTitle{icemelt}
}
\description{
\packageDescription{icemelt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{icemelt}
\packageIndices{icemelt}
}
\author{
\packageAuthor{icemelt}

Maintainer: \packageMaintainer{icemelt}
}
\references{
Mandal, S., Wang, S. and Sinha, S. (2019+). Analysis of Linear Transformation Models with Covariate Measurement Error and Interval Censoring. (accepted, Statistics In Medicine)
L'ecuyer, P., (1988). Efficient and portable combined random number generators. Communications of the ACM, 31(6), pp.742-751.
}
\keyword{ package }
